# Optimize character to index conversion for a 4% decrease on CONNL 2003