--- a/flair/data.py
+++ b/flair/data.py
@@ -1,4 +1,5 @@
 from abc import abstractmethod
+from operator import itemgetter
 from typing import List, Dict, Union, Callable
 
 import torch, flair
@@ -27,7 +28,7 @@ class Dictionary:
         # init dictionaries
         self.item2idx: Dict[str, int] = {}
         self.idx2item: List[str] = []
-        self.item2idx_not_encoded: Dict[str, int] = {}
+        self.item2idx_not_encoded: Dict[str, int] = defaultdict(int)
         self.idx2item_not_encoded: List[str] = []
         self.multi_label: bool = False
         # in order to deal with unknown tokens, add <unk>
@@ -55,7 +56,12 @@ class Dictionary:
         """
         if not hasattr(self, "item2idx_not_encoded"):
             self.decode_utf_internal()
-        return self.item2idx_not_encoded.get(item, 0)
+        return self.item2idx_not_encoded[item]
+
+    def get_idx_for_items(self, items: List[str]) -> List[int]:
+        if not hasattr(self, "item2idx_not_encoded"):
+            self.decode_utf_internal()
+        return itemgetter(*items)(self.item2idx_not_encoded)
 
     def get_items(self) -> List[str]:
         if not hasattr(self, "item2idx_not_encoded"):
@@ -73,8 +79,10 @@ class Dictionary:
         return self.idx2item_not_encoded[idx]
 
     def decode_utf_internal(self):
-        self.idx2item_not_encoded = [item.decode("UTF-8") for item in self.idx2item]
-        self.item2idx_not_encoded = dict([(key.decode("UTF-8"), value) for key, value in self.item2idx.items()])
+        if not hasattr(self, "item2idx_not_encoded"):
+            self.idx2item_not_encoded = [item.decode("UTF-8") for item in self.idx2item]
+            d = dict([(key.decode("UTF-8"), value) for key, value in self.item2idx.items()])
+            self.item2idx_not_encoded = defaultdict(int, d)
 
     def save(self, savefile):
         if not hasattr(self, "item2idx_not_encoded"):