--- a/flair/data.py
+++ b/flair/data.py
@@ -28,7 +28,7 @@ class Dictionary:
         # init dictionaries
         self.item2idx: Dict[str, int] = {}
         self.idx2item: List[str] = []
-        self.item2idx_not_encoded: Dict[str, int] = defaultdict(int)
+        self.item2idx_not_encoded: Dict[str, int] = defaultdict(lambda: 0)
         self.idx2item_not_encoded: List[str] = []
         self.multi_label: bool = False
         # in order to deal with unknown tokens, add <unk>
@@ -59,9 +59,19 @@ class Dictionary:
         return self.item2idx_not_encoded[item]
 
     def get_idx_for_items(self, items: List[str]) -> List[int]:
+        """
+        returns the IDs for each item of the list of string, otherwise 0 if not found
+        :param items: List of string for which IDs are requested
+        :return: List of ID of strings
+        """
         if not hasattr(self, "item2idx_not_encoded"):
             self.decode_utf_internal()
-        return itemgetter(*items)(self.item2idx_not_encoded)
+        if not items:
+            return []
+        results = itemgetter(*items)(self.item2idx_not_encoded)
+        if isinstance(results, int):
+            return [results]
+        return list(results)
 
     def get_items(self) -> List[str]:
         if not hasattr(self, "item2idx_not_encoded"):
@@ -82,7 +92,7 @@ class Dictionary:
         if not hasattr(self, "item2idx_not_encoded"):
             self.idx2item_not_encoded = [item.decode("UTF-8") for item in self.idx2item]
             d = dict([(key.decode("UTF-8"), value) for key, value in self.item2idx.items()])
-            self.item2idx_not_encoded = defaultdict(int, d)
+            self.item2idx_not_encoded = defaultdict(lambda: 0, d)
 
     def save(self, savefile):
         if not hasattr(self, "item2idx_not_encoded"):
@@ -91,7 +101,7 @@ class Dictionary:
         import pickle
 
         with open(file=savefile, mode="wb") as f:
-            mappings = {"idx2item": self.idx2item_not_encoded, "item2idx": self.item2idx_not_encoded, "UTF-8": True}
+            mappings = {"idx2item": self.idx2item_not_encoded, "item2idx": dict(self.item2idx_not_encoded), "UTF-8": True}
             pickle.dump(mappings, f)
 
     @classmethod