Optimize the way padding is done by reducing the number of dictionary call.
Remove the conversion to UTF-8, and take care of old models in old format.
In the future, the conversion may be removed safely.

close https://github.com/zalandoresearch/flair/issues/1133
@alanakbik finally I did both, change in dictionary deserialization and change inside the class itself.

CONNL 2003 from 24s to 23s (-4%)
French dataset from 12s to 11s (-8%)