@pommedeterresautee 2019-09-23
There was strange issues because of dictionary conversion, so I finally made a small version of the optimization, keeping things like they are :-( May be to investigate in the future