@alanakbik 2019-09-23
Ah ok I was just measuring speed differences and finding they are roughly the same. Is this because of the last changes?