@pommedeterresautee 2019-09-23
Here I still have 23s and 11s for CONNL and French dataset. May be it's a less than 1 sec change but with rounding it appears bigger?